requirementsHenil.txt
asgiref==3.4.1
asttokens==2.2.1
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.1.7
Django==3.2.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extensions==3.2.1
django-lifecycle==1.0.0
django-phonenumber-field==7.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-writable-nested==0.7.0
drf-yasg==1.21.5
executing==1.2.0
idna==2.10
inflection==0.5.1
ipython==8.12.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.14
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.20
Pygments==2.15.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
razorpay==1.3.0
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.22

ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.1
stack-data==0.6.2
traitlets==5.9.0
twilio==7.15.3
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.6
urlman==2.0.1
wcwidth==0.2.6
requirementsHenil.txt